requirements.txt
aiohttp>=3.8.0
dnspython>=2.1.0